Revert "Move DevAddrCacheInfo and IoTHubDeviceInfo to LoraTools"
This reverts commit 91e210d9c24c4894d11313ba7b1ecbc254446466.
